… Sources/SignalProducer.swift
… Sources/TupleExtensions.swift