Preceded by https://github.com/ReactiveCocoa/ReactiveCocoa/pull/2521.
Fixes #156.

1. One lock and one `Signal` regardless of the arity or the number of signals.
1. Stack friendly.
1. Human maintainable IMO.

| Test item, 50000 iterations | `master` | This PR | Delta | Speedup |
| ---- | ---- | ---- | --- | --- |
| Tenary `combineLatest`, `Signal` | 0.370 sec | 0.205 sec | -18% | 1.80x |
| Ternary `combineLatest`, `SignalProducer` | 0.589 sec | 0.313 sec | -47% | 1.88x |
| 6-ary `combineLatest`, `Signal` | 1.733 sec | 0.440 sec | -75% | 3.94x |
| 6-ary `combineLatest`, `SignalProducer` | 2.541 sec | 0.651 sec | -74% | 3.90x |
| Blowing Main Stack (1 iteration) | 4950 | 7450 | N/A | N/A |

### Note
`combineLatest` and `zip` adds signals in a chain to a `Singal.AggregateBuilder`, which constructs a `Signal` when the chain ends with `make()`. `Singal.AggregateBuilder` encapsulates all the type erasure & recovery.

In https://github.com/ReactiveCocoa/ReactiveCocoa/pull/2521, a primary dissent was that it introduced risks that the compiler cannot statically validate. This PR does not, and cannot mitigate, such risk with the performance goal due to the lack of variadic generics.

Note: Xcode 8.3.2, i5 6360U, Whole Module Optimisation. [Benchmark Source (Gist)](https://gist.github.com/andersio/5eaef64adccc4af9f687abc8894cc5de).