@andersio 2017-05-02
@mdiep https://github.com/ReactiveCocoa/ReactiveSwift/pull/345/commits/6ee1fe723f77e0f3c4978585c172ead1ef50045b probably still doesn't do what you want. But I don't think nested calls would make it any more manageable.