@andersio 2017-05-02
This seems working, but we would have to pull back `TupleExtensions.swift`. It also takes `2n - 3` calls for a n-arity operator to pack its results.

```swift
struct Builder<Composite, E> {
    func append<U, NewComposite>(_ signal: S<U, E>, _ pack: @escaping (Composite, Any) -> NewComposite) -> Builder<NewComposite, E> {
        let transform = self.transform
        let newTransform = { pack(firstTransform($0), $0[0]) }
    }
}

let _: Builder<(A, B, C, D, E), E> = Builder<(A, B), E>(a, b)
    .append(c, repack)
    .append(d, repack)
    .append(e, repack)
```