@mdiep 2017-05-03
This is what I was thinking:

```swift
extension Signal {
  func erased() -> Signal<Any, Error> { … }
}

class Combiner<Strategy> {
  var startHandlers: […]

  init<Composite>(
    signals: [Signal<Any, Error>],
    _ transform: @escaping (ContiguousArray<Any>) -> Composite
  ) -> Signal<Composite, Error> {
    // the code that was in make(count:_:), but also the code from add(_:)
  }

  convenience init<A, B>(
    _ a: Signal<A, Error>,
    _ b: Signal<B, Error>
  ) -> Signal<(A, B>), Error> {
    self.init([a.erased(), b.erased()]) { ($0[0] as! A, $0[1] as! B) }
  }

  …
}
```