@andersio 2017-05-03
I feel like https://github.com/ReactiveCocoa/ReactiveSwift/pull/345/commits/6ee1fe723f77e0f3c4978585c172ead1ef50045b is pretty close, except for not needing an erased version of signals.